+++ substratevm/src/com.oracle.svm.truffle/src/com/oracle/svm/truffle/isolated/IsolatedCompilableTruffleAST.java
@@ -27,7 +27,6 @@ package com.oracle.svm.truffle.isolated;
 import java.util.function.Supplier;
 
 import org.graalvm.compiler.truffle.common.CompilableTruffleAST;
-import org.graalvm.compiler.truffle.common.OptimizedAssumptionDependency;
 import org.graalvm.compiler.truffle.common.TruffleCallNode;
 import org.graalvm.nativeimage.PinnedObject;
 import org.graalvm.nativeimage.c.function.CEntryPoint;
@@ -136,18 +135,18 @@ final class IsolatedCompilableTruffleAST extends IsolatedObjectProxy<SubstrateCo
     }
 
     @Override
-    public SubstrateInstalledCode getSubstrateInstalledCode() {
+    public JavaConstant getValidRootAssumptionConstant() {
+        return new IsolatedObjectConstant(getValidRootAssumption0(IsolatedCompileContext.get().getClient(), handle), false);
+    }
+
+    @Override
+    public SubstrateInstalledCode createSubstrateInstalledCode() {
         throw VMError.shouldNotReachHere("Must not be called during isolated compilation");
     }
 
     @Override
-    public OptimizedAssumptionDependency getDependency() {
-        throw VMError.shouldNotReachHere("Must not be called during isolated compilation");
-    }
-
-    @Override
-    public InstalledCode createInstalledCode() {
-        return new IsolatedCodeInstallBridge(handle, handle);
+    public InstalledCode createPreliminaryInstalledCode() {
+        return new IsolatedCodeInstallBridge(handle);
     }
 
     @Override
@@ -269,6 +268,15 @@ final class IsolatedCompilableTruffleAST extends IsolatedObjectProxy<SubstrateCo
     private static ClientHandle<Assumption> getNodeRewritingAssumption0(@SuppressWarnings("unused") ClientIsolateThread client, ClientHandle<SubstrateCompilableTruffleAST> handle) {
         CompilableTruffleAST ast = IsolatedCompileClient.get().unhand(handle);
         JavaConstant assumptionConstant = ast.getNodeRewritingAssumptionConstant();
+        Assumption assumption = KnownIntrinsics.convertUnknownValue(SubstrateObjectConstant.asObject(assumptionConstant), Assumption.class);
+        return IsolatedCompileClient.get().hand(assumption);
+    }
+
+    @CEntryPoint
+    @CEntryPointOptions(include = CEntryPointOptions.NotIncludedAutomatically.class, publishAs = CEntryPointOptions.Publish.NotPublished)
+    private static ClientHandle<Assumption> getValidRootAssumption0(@SuppressWarnings("unused") ClientIsolateThread client, ClientHandle<SubstrateCompilableTruffleAST> handle) {
+        CompilableTruffleAST ast = IsolatedCompileClient.get().unhand(handle);
+        JavaConstant assumptionConstant = ast.getValidRootAssumptionConstant();
         Assumption assumption = KnownIntrinsics.convertUnknownValue(SubstrateObjectConstant.asObject(assumptionConstant), Assumption.class);
         return IsolatedCompileClient.get().hand(assumption);
     }